+++ packages/framework/CHANGELOG.md
@@ -1,3 +1,18 @@
+## 1.0.2
+
+### Breaking change ❗️
+
+- Framework API change:
+  - `analyze(entry, files)` -> `analyze(files)`
+  - `compile()` -> `compile(entry)`
+
+### Bug fix 🐛
+
+- `framework/react`
+  - remove react specific version
+- `framework/html`
+  - fix entry name bug: the entry name must end with `.html` to be recognized
+
 ## 1.0.1
 
 ### Bug fix 🐛